criação de jogo / remoção do main
